remove pass_manager.py create separate classes
